dev: minor refactor
