+++ cmd/create/date.go
@@ -1,4 +1,4 @@
-package cmd
+package create
 
 type _Date string
